requirements.txt
# LangChain Core
LangChain
langchain-Core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic


# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Huggingface Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variables Managements
python-dotenv


# Machine Learning Utilities
numpy
scikit-learn